fix: harden hosted catalog validation
