+++ packages/ai/src/index.ts
@@ -4,6 +4,7 @@ export { Type } from "typebox";
 export * from "./api-registry.js";
 export * from "./env-api-keys.js";
 export * from "./log.js";
+export * from "./model-catalog.js";
 export * from "./models.js";
 export type { BedrockOptions, BedrockThinkingDisplay } from "./providers/amazon-bedrock.js";
 export type { AnthropicEffort, AnthropicOptions, AnthropicThinkingDisplay } from "./providers/anthropic.js";
